python/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
python-pptx==1.0.2
pydantic==2.10.6
python-multipart==0.0.20
pyttsx3==2.98
pywin32>=308; platform_system == "Windows"
